requirements.txt
fastapi==0.111.0
SQLAlchemy==2.0.30
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
alembic==1.13.1